address comments and minor refactor and translation fix
