+++ libs/application/templates/accident-notification/src/lib/messages/inReview.ts
@@ -129,13 +129,13 @@ export const inReview = {
   sjukratrygging: defineMessages({
     title: {
       id: 'an.application:inReview.sjukratrygging.title',
-      defaultMessage: 'Afstaða til bótastöðu',
+      defaultMessage: 'Afstaða til bótaskyldu',
       description: 'Title for sjukratrygging section in review',
     },
     summary: {
       id: 'an.application:inReview.sjukratrygging.summary',
       defaultMessage:
-        'Þegar öll nauðsynleg gögn hafa borist, fer Sjúkratryggingar Íslands yfir umsókn og er þá afstaða tekin til bótaskyldu.',
+        'Þegar öll nauðsynleg gögn hafa borist, fara Sjúkratryggingar Íslands yfir umsókn og er þá afstaða tekin til bótaskyldu.',
       description: 'Summary for sjukratrygging section in review',
     },
   }),
@@ -175,7 +175,7 @@ export const inReview = {
       description: {
         id: 'an.application:inReview.action.representative.description',
         defaultMessage:
-          'Það er nauðsynlegt að forsvarsmaður fari yfir tilkynningu áður en Sjúkratryggingar Íslands getur tekið afstöðu til bótaskyldu.',
+          'Það er nauðsynlegt að forsvarsmaður eða hinn slasaði fari yfir tilkynningu áður en Sjúkratryggingar Íslands getur tekið afstöðu til bótaskyldu.',
         description: 'Description for action message in representative section',
       },
       actionButtonTitle: {
