+++ libs/application/templates/data-protection-complaint/src/fields/index.ts
@@ -9,4 +9,4 @@ export { ComplaintDescription } from './ComplaintDescription/ComplaintDescriptio
 export { ComplaintDocumentHeading } from './ComplaintDocumentHeading'
 export { AgreementDescription } from './AgreementDescription'
 export { CommissionDocument } from './CommissionDocument'
-export { ReasonsForComplaint } from './ConditionallyRequiredInput'
+export { ReasonsForComplaint } from './ReasonsForComplaint'
